requirements.txt
fastapi
uvicorn[standard]
gspread
google-auth
httpx
python-dotenv
pytz
